pyproject.toml
[project]
name = "backend-rag-ia"
version = "1.0.0"
description = "Backend com RAG e sistema multiagente"
authors = [
    {name = "Flow", email = "flow@example.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
    "rich>=13.7.0",
    "pydantic>=2.5.2",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6"
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"
